changed percentage to a double

+++ src/Board.java
@@ -2,10 +2,10 @@
 public class Board {
 	
 	private int size;
-	private int bombPer;
+	private double bombPer;
 	private int[][] gameBoard;
 	
-	public Board(int s, int b) {
+	public Board(int s, double b) {
 		this.size = s;
 		this.bombPer = b;
 		gameBoard = new int[s][s];
